+++ test/function/matrix/filter.test.js
@@ -18,6 +18,53 @@ describe('filter', function() {
     assert.deepEqual(math.filter(math.matrix([6, -2, -1, 4, 3]), isPositive), math.matrix([6, 4, 3]));
   });
 
+  it('should invoke callback with parameters value, index, obj', function() {
+    var arr = [1,2,3];
+    var log = [];
+
+    math.filter(arr, function (value, index, obj) {
+      log.push([value, index, obj === arr]);
+      return true;
+    }).valueOf();
+
+    assert.deepEqual(log, [
+      [1, [0], true ],
+      [2, [1], true ],
+      [3, [2], true ]
+    ]);
+
+  });
+
+  it('should invoke a typed function with correct number of arguments (1)', function() {
+    var output = [];
+    math.filter([1,2,3], math.typed('callback', {
+      'number': function (value) {
+        output.push(value + 2)
+      }
+    }));
+    assert.deepEqual(output, [3,4,5]);
+  });
+
+  it('should invoke a typed function with correct number of arguments (2)', function() {
+    var output = [];
+    math.filter([1,2,3], math.typed('callback', {
+      'number, Array': function (value, index) {
+        output.push(value + 2)
+      }
+    }));
+    assert.deepEqual(output, [3,4,5]);
+  });
+
+  it('should invoke a typed function with correct number of arguments (3)', function() {
+    var output = [];
+    math.filter([1,2,3], math.typed('callback', {
+      'number, Array, Array': function (value, index, array) {
+        output.push(value + 2)
+      }
+    }));
+    assert.deepEqual(output, [3,4,5]);
+  });
+
   it('should filter an array with a regexp', function() {
     assert.deepEqual(math.filter(["23", "foo", "100", "55", "bar"], /[0-9]+/), ["23", "100", "55"]);
   });